requirements.txt
groq>=0.9.0
radon>=6.0.1
bandit>=1.7.5
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
